refactor(bevy_gantz_collab): split lib.rs into ui, session and sync modules
Pure code motion along the crate's three seams: ui.rs (dispatchers, the
settings-tab provider, the display mirror, presence), session.rs (share/
join/leave lifecycle and head bookkeeping), sync.rs (the fetch/converge
plane: event drain, want/fetch loop, per-name convergence, announcing,
SyncCtx). The plugin, resources and events stay in lib.rs, which
re-exports the public items so external paths are unchanged.
